# Optimizing the way we calculate our in-memory pool size