--- a/lib/optimization/event_test.go
+++ b/lib/optimization/event_test.go
@@ -287,7 +287,7 @@ func TestTableData_ShouldFlushRowSize(t *testing.T) {
 
 	// Insert 3 rows and confirm that we need to flush.
 	td := NewTableData(nil, nil, kafkalib.TopicConfig{}, "foo")
-	for i := 0; i < 85; i++ {
+	for i := 0; i < 100; i++ {
 		shouldFlush, flushReason := td.ShouldFlush(ctx)
 		assert.False(t, shouldFlush)
 		assert.Empty(t, flushReason)
@@ -313,6 +313,8 @@ func TestTableData_ShouldFlushRowSize(t *testing.T) {
 			"123": "9222213213j1i31j3k21j321k3j1k31jk31213123213213121322j31k2",
 		},
 	}, false)
+
+	fmt.Println("####", td.approxSize)
 
 	shouldFlush, flushReason := td.ShouldFlush(ctx)
 	assert.True(t, shouldFlush)