# TL;DR
![giphy](https://github.com/artie-labs/transfer/assets/4412200/bc02bc60-b7e2-49ff-a6cf-9e7f2c889ae3)

## Problem

Previously, we were encoding the values to be string and then calculating the number of bytes to approximate how large each row is.

For super wide tables, it was pretty slow and could take up to 20ms each time it was invoked.

## Change

Now, we are iterating over the row and then calculating each row by first inspecting the primitive type and if it's indeed a string, just count the number of bytes as opposed to serializing it out again.

## Results

```
BEFORE
cpu: VirtualApple @ 2.50GHz
BenchmarkGetApproxSize_TallTable-12    	    1604	   7524981 ns/op
BenchmarkGetApproxSize_WideTable-12    	     214	  54952219 ns/op

BenchmarkTableData_ApproxSize_TallTable-12    	 7741323	      1654 ns/op
BenchmarkTableData_ApproxSize_WideTable-12    	 1000000	     11223 ns/op

-----
AFTER
BenchmarkGetApproxSize_TallTable-12    	   35970	    333518 ns/op
BenchmarkGetApproxSize_WideTable-12    	     745	  15972217 ns/op

BenchmarkTableData_ApproxSize_TallTable-12    	19568608	       743.2 ns/op
BenchmarkTableData_ApproxSize_WideTable-12    	 2203387	      6018 ns/op
```

* After is 1.8x faster for wide tables (the bottom 2 tests)
* After is 2.2x faster for tall tables